+++ prisma/schema.prisma
@@ -14,11 +14,48 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
+model Account {
+  id                 String    @id @default(auto()) @map("_id") @db.ObjectId
+  userId             String
+  providerType       String?
+  type               String?
+  provider           String
+  providerAccountId  String
+  refreshToken       String?
+  access_token       String?
+  token_type         String?
+  scope              String?
+  accessTokenExpires DateTime?
+  createdAt          DateTime  @default(now())
+  updatedAt          DateTime  @updatedAt
+  user               User      @relation(fields: [userId], references: [id])
+
+  @@unique([provider, providerAccountId])
+}
+
+model Session {
+  id           String   @id @default(auto()) @map("_id") @db.ObjectId
+  userId       String
+  expires      DateTime
+  sessionToken String   @unique
+  accessToken  String?   @unique
+  createdAt    DateTime @default(now())
+  updatedAt    DateTime @updatedAt
+  user         User     @relation(fields: [userId], references: [id])
+}
+
 model User {
-  id      String   @id @default(auto()) @map("_id") @db.ObjectId
-  email   String   @unique
-  name    String?
+  id            String    @id @default(auto()) @map("_id") @db.ObjectId
+  name          String?
+  email         String?   @unique
+  emailVerified DateTime?
+  image         String?
+  createdAt     DateTime  @default(now())
+  updatedAt     DateTime  @updatedAt
+  accounts      Account[]
+  sessions      Session[]
   entries    Entry[]
+
 }
 
 type Entry {
@@ -32,6 +69,7 @@ type Day {
   earnings String
   availableBalance String
   date String
+  status String
 }
 
 type Week {
@@ -39,12 +77,15 @@ type Week {
   earnings String
   availableBalance String
   date String
+  status String
+  days Day[]
 }
 
 type Task {
   category Category
   area Areas
-  status Boolean
+  status String
+  cadence String
 }
 
 enum Areas {
