+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,7 +1,11 @@
 import NextAuth from "next-auth"
+import { PrismaAdapter } from "@next-auth/prisma-adapter"
 import GithubProvider from "next-auth/providers/github"
+import prisma from "@/lib/prisma";
+
 export const authOptions = {
   // Configure one or more authentication providers
+  secret: process.env.NEXTAUTH_SECRET,
   providers: [
     GithubProvider({
       clientId: process.env.GITHUB_ID,
@@ -13,7 +17,8 @@ export const authOptions = {
     async redirect({ url, baseUrl }) {
       return '/app/day'
     }
-  }
+  },
+  adapter: PrismaAdapter(prisma)
 }
 const handler = NextAuth(authOptions)
